# contracts/openapi/control-plane.v1.json
{
  "components": {
    "schemas": {
...
      "DouyinCandidateDisposition": {
        "enum": [
          "eligible",
          "duplicate_in_task",
          "duplicate_in_history",
          "blacklisted"
        ],
        "title": "DouyinCandidateDisposition",
        "type": "string"
      },
      "DouyinCandidateSource": {
        "enum": [
          "general_search_author"
        ],
        "title": "DouyinCandidateSource",
        "type": "string"
      },
      "DouyinSearchExposureAction": {
        "enum": [
          "browse",
          "comment",
          "direct_message"
        ],
        "title": "DouyinSearchExposureAction",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TaskCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/DouyinSearchExposureAction"
          },
          "finalConfirmationRequired": {
            "const": true,
            "title": "Finalconfirmationrequired",
            "type": "boolean"
          },
          "maximumIntervalSeconds": {
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Maximumintervalseconds",
            "type": "integer"
          },
          "messageTemplate": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Messagetemplate"
          },
          "minimumIntervalSeconds": {
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Minimumintervalseconds",
            "type": "integer"
          },
          "previewRequired": {
            "const": true,
            "title": "Previewrequired",
            "type": "boolean"
          },
          "searchKeyword": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Searchkeyword",
            "type": "string"
          },
          "targetLimit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Targetlimit",
            "type": "integer"
          },
          "template": {
            "const": "douyin.search_exposure.v1",
            "title": "Template",
            "type": "string"
          }
        },
        "required": [
          "template",
          "searchKeyword",
          "action",
          "messageTemplate",
          "targetLimit",
          "minimumIntervalSeconds",
          "maximumIntervalSeconds",
          "previewRequired",
          "finalConfirmationRequired"
        ],
        "title": "TaskCreateRequest",
        "type": "object"
      },
...
      "TaskListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "TaskListResponse",
        "type": "object"
      },
      "TaskResponse": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "lastEventSequence": {
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Lasteventsequence",
            "type": "integer"
          },
          "revision": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "taskId": {
            "title": "Taskid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "taskId",
          "status",
          "revision",
          "lastEventSequence",
          "createdAt",
          "updatedAt"
        ],
        "title": "TaskResponse",
        "type": "object"
      },
      "TaskStatus": {
        "description": "Persisted and public lifecycle states for one task.",
        "enum": [
          "draft",
          "validating",
          "awaiting_device",
          "awaiting_platform_login",
          "discovering_targets",
          "awaiting_confirmation",
          "queued",
          "running",
          "paused",
          "awaiting_human",
          "cancelling",
          "succeeded",
          "partially_succeeded",
          "failed",
          "cancelled",
          "outcome_uncertain"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
      "TaskTargetConfirmationRequest": {
        "additionalProperties": false,
        "properties": {
          "expectedTaskRevision": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Expectedtaskrevision",
            "type": "integer"
          },
          "pageRevision": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Pagerevision",
            "type": "integer"
          }
        },
        "required": [
          "pageRevision",
          "expectedTaskRevision"
        ],
        "title": "TaskTargetConfirmationRequest",
        "type": "object"
      },
      "TaskTargetExclusionsRequest": {
        "additionalProperties": false,
        "properties": {
          "excludedTargetIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Excludedtargetids",
            "type": "array"
          },
          "expectedTaskRevision": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Expectedtaskrevision",
            "type": "integer"
          },
          "pageRevision": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Pagerevision",
            "type": "integer"
          }
        },
        "required": [
          "pageRevision",
          "expectedTaskRevision",
          "excludedTargetIds"
        ],
        "title": "TaskTargetExclusionsRequest",
        "type": "object"
      },
      "TaskTargetPreviewItemResponse": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "disposition": {
            "$ref": "#/components/schemas/DouyinCandidateDisposition"
          },
          "ordinal": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Ordinal",
            "type": "integer"
          },
          "publicHandle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publichandle"
          },
          "selected": {
            "title": "Selected",
            "type": "boolean"
          },
          "source": {
            "$ref": "#/components/schemas/DouyinCandidateSource"
          },
          "targetId": {
            "title": "Targetid",
            "type": "string"
          },
          "userExcluded": {
            "title": "Userexcluded",
            "type": "boolean"
          }
        },
        "required": [
          "targetId",
          "ordinal",
          "displayName",
          "publicHandle",
          "source",
          "disposition",
          "userExcluded",
          "selected"
        ],
        "title": "TaskTargetPreviewItemResponse",
        "type": "object"
      },
      "TaskTargetPreviewResponse": {
        "additionalProperties": false,
        "properties": {
          "confirmed": {
            "title": "Confirmed",
            "type": "boolean"
          },
          "confirmedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmedat"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskTargetPreviewItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "lastEventSequence": {
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Lasteventsequence",
            "type": "integer"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "pageRevision": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Pagerevision",
            "type": "integer"
          },
          "selectedTargetCount": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Selectedtargetcount",
            "type": "integer"
          },
          "taskId": {
            "title": "Taskid",
            "type": "string"
          },
          "taskRevision": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Taskrevision",
            "type": "integer"
          },
          "taskStatus": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "userExcludedTargetCount": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Userexcludedtargetcount",
            "type": "integer"
          }
        },
        "required": [
          "taskId",
          "taskStatus",
          "taskRevision",
          "lastEventSequence",
          "pageRevision",
          "selectedTargetCount",
          "userExcludedTargetCount",
          "confirmed",
          "confirmedAt",
          "items",
          "nextCursor"
        ],
        "title": "TaskTargetPreviewResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AppSession": {
        "scheme": "bearer",
        "type": "http"
      },
      "DemoBootstrap": {
        "scheme": "bearer",
        "type": "http"
      },
      "DeviceCredential": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "automation-tool Control Plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "operationId": "listTasks",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AppSession": []
          }
        ],
        "summary": "List Tasks",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "createTask",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AppSession": []
          }
        ],
        "summary": "Create Task",
        "tags": [
          "tasks"
        ]
      }
    },
...
    "/api/v1/tasks/{task_id}/target-preview": {
      "get": {
        "operationId": "getTaskTargetPreview",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTargetPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AppSession": []
          }
        ],
        "summary": "Get Task Target Preview",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/target-preview/confirmations": {
      "post": {
        "operationId": "confirmTaskTargetPreview",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskTargetConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTargetPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AppSession": []
          }
        ],
        "summary": "Confirm Task Target Preview",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/target-preview/exclusions": {
      "put": {
        "operationId": "replaceTaskTargetExclusions",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskTargetExclusionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTargetPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AppSession": []
          }
        ],
        "summary": "Replace Task Target Exclusions",
        "tags": [
          "tasks"
        ]
      }
    },
...
  }
}